+++ forui/lib/src/widgets/calendar/shared/entry.dart
@@ -212,21 +212,21 @@ class _Content extends StatelessWidget {
 final class FCalendarEntryStyle with Diagnosticable, _$FCalendarEntryStyleFunctions {
   /// The day's background color.
   ///
-  /// {@macro forui.foundation.tappable.builder}
+  /// {@macro forui.foundation.FTappable.builder}
   /// * [WidgetState.selected]
   @override
   final FWidgetStateMap<Color> backgroundColor;
 
   /// The border.
   ///
-  /// {@macro forui.foundation.tappable.builder}
+  /// {@macro forui.foundation.FTappable.builder}
   /// * [WidgetState.selected]
   @override
   final FWidgetStateMap<Color?> borderColor;
 
   /// The day's text style.
   ///
-  /// {@macro forui.foundation.tappable.builder}
+  /// {@macro forui.foundation.FTappable.builder}
   /// * [WidgetState.selected]
   @override
   final FWidgetStateMap<TextStyle> textStyle;